Move player's files into own directories
